requirements.txt
fiftyone>=0.22.0
transformers>=4.30.0
torch>=2.0.0
einops 
timm
accelerate
sentence-transformers>=2.2.2
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn
certifi
requests
plotly>=5.13.0
clip @ git+https://github.com/openai/CLIP.git